backend/requirements.txt
amqp==5.0.6
apns2==0.7.2
arabic-reshaper==2.1.3
asn1crypto==1.5.1
attrs==21.2.0
billiard==3.6.4.0
Brotli==1.0.9
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==35.0.0
cssselect2==0.5.0
dj-database-url==0.5.0
Django==2.2.12
django-configurations==2.2
django-cors-headers==3.10.1
django-filter==21.1
django-memcache-status==2.2
django-model-utils==4.2.0
django-modeladmin-reorder==0.3.1
django-phonenumber-field==5.2.0
django-push-notifications==2.0.0
django-tinymce==3.3.0
django-unique-upload==0.2.1
django-unused-media==0.2.2
djangorestframework==3.12.4
djangorestframework-gis==0.17
drf-spectacular==0.20.2
et-xmlfile==1.1.0
factory-boy==3.2.0
Faker==9.5.2
flower==1.0.0
fonttools==4.32.0
future==0.18.2
geographiclib==1.52
geopy==2.2.0
gsconfig-py3==1.0.8
h2==2.6.2
hpack==3.0.0
html5lib==1.1
http-ece==1.1.0
humanize==3.12.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.3
inflection==0.5.1
Jinja2==3.1.1
jsonschema==4.1.2
kombu==5.1.0
lxml==4.8.0
Markdown==3.3.4
MarkupSafe==2.1.1
numpy==1.21.5
openpyxl==3.0.9
oscrypto==1.3.0
pandas==1.3.5
phonenumbers==8.12.35
Pillow==8.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.21
py-vapid==1.8.2
pycparser==2.20
pydyf==0.1.2
pyfcm==1.5.4
pyHanko==0.12.1
pyhanko-certvalidator==0.19.5
PyJWT==1.7.1
PyPDF3==1.0.6
pyphen==0.12.0
pyrsistent==0.18.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.5
python-docx==0.8.11
python-memcached==1.59
python-slugify==5.0.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywebpush==1.14.0
PyYAML==6.0
qrcode==7.3.1
redis==3.5.3
reportlab==3.6.9
requests==2.26.0
sentry-sdk==1.4.3
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
text-unidecode==1.3
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
weasyprint==54.3
webencodings==0.5.1
websockets==10.1
xhtml2pdf==0.2.7
xlrd==2.0.1
XlsxWriter==3.0.3
zopfli==0.2.1
